Fixing Win 64 conversion warnings

+++ trunk/src/segyprogs/segy_io.c
@@ -4,7 +4,7 @@
 
 
 /*
- *	$Id: segy_io.c,v 1.3 2008-03-24 08:58:33 guru Exp $
+ *	$Id: segy_io.c,v 1.4 2011-03-05 21:24:29 guru Exp $
  * segy_io.c:  A suite of functions to cover reading SEGY header variables.
  *
  * modified by T. Henstock from the PASSCAL software suite.
@@ -41,7 +41,7 @@ int get_segy_reelhd (fileptr, reelhead)
 	FILE *fileptr;
 	char *reelhead;
 {
-	int igot;
+	size_t igot;
 
 
         if ((igot = fread (reelhead, (size_t)3200, (size_t)1, fileptr)) != 1) {
@@ -58,7 +58,7 @@ int get_segy_binhd (fileptr, binhead)
 	FILE *fileptr;
 	SEGYREEL *binhead;
 {		
-	int igot;
+	size_t igot;
 
 
 	if ((igot = fread (binhead, (size_t)400, (size_t)1, fileptr)) !=1) {
